Create CoinList files

+++ app/src/main/java/ru/dubr/traineetestandroid/data/network/dto/CoinDto.kt
@@ -18,13 +18,13 @@ data class CoinDto(
     @SerializedName("market_cap")
     val marketCap: Long,
     @SerializedName("market_cap_rank")
-    val marketCapRank: Int,
+    val marketCapRank: Double,
     @SerializedName("fully_diluted_valuation")
     val fullyDilutedValuation: Long,
     @SerializedName("total_volume")
     val totalVolume: Long,
     @SerializedName("high_24h")
-    val high24h: Int,
+    val high24h: Double,
     @SerializedName("low_24h")
     val low24h: Double,
     @SerializedName("price_change_24h")
@@ -36,13 +36,13 @@ data class CoinDto(
     @SerializedName("market_cap_change_percentage_24h")
     val marketCapChangePercentage24h: Double,
     @SerializedName("circulating_supply")
-    val circulatingSupply: Int,
+    val circulatingSupply: Double,
     @SerializedName("total_supply")
-    val totalSupply: Int,
+    val totalSupply: Double,
     @SerializedName("max_supply")
-    val maxSupply: Int,
+    val maxSupply: Double,
     @SerializedName("ath")
-    val ath: Int,
+    val ath: Double,
     @SerializedName("ath_change_percentage")
     val athChangePercentage: Double,
     @SerializedName("ath_date")
